+++ src/polymer/docs.ts
@@ -87,7 +87,7 @@ export function annotateEvent(annotation: jsdoc.Annotation): ScannedEvent {
   }
   const scannedEvent: ScannedEvent = {
     name: name,
-    description: (eventTag && eventTag.description) || annotation.description,
+    description: annotation.description || (eventTag && eventTag.description) || undefined,
     jsdoc: annotation,
     sourceRange: undefined,
     astNode: null,